storage: improve the setting of br.Timestamp
Before this patch, evaluateBatch() had special code to deal with
transactions committed at a different timestamp than txn.ReadTimestamp.
This was useful when the EndTxn request decided that it can commit
despite a bumped write timestamp.
This patch simplifies this code by having EndTxn update the read
timestamp. This way, evaluateBatch() no longer needs to consider the
transaction's disposition. The burden is moved to EndTxn, which is
already in this dirty business.

The change is also in the spirit of how timestamps are supposed to work.
For transactional requests, code should generally look at
ba.Txn.ReadTimestamp, not at ba.Timestamp. However, evaluateBatch() was
using ba.Timestamp around the affected code.

Release note: None